# api/docs/swagger.yaml
basePath: /
definitions:
...
  github_com_erhan_falcon_api_internal_models.Report:
    properties:
      created_at:
        type: string
      cvss:
        type: string
      description_md:
        type: string
      endpoint_id:
        $ref: '#/definitions/sql.NullInt64'
      external_url:
        type: string
      host_id:
        $ref: '#/definitions/sql.NullInt64'
      id:
        type: integer
      poc:
        type: string
      program_id:
        type: integer
      reward_amount:
        $ref: '#/definitions/sql.NullFloat64'
      reward_currency:
        type: string
      service_id:
        $ref: '#/definitions/sql.NullInt64'
      severity:
        type: string
      status:
        type: string
      target:
        type: string
      title:
        type: string
      updated_at:
        type: string
      url:
        type: string
      vrt:
        type: string
    type: object
...
  internal_handlers.errorResp:
    properties:
      error:
        type: string
    type: object
...
  internal_handlers.pageResp:
    properties:
      items:
        items: {}
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
...
  internal_handlers.reportReq:
    properties:
      cvss:
        type: string
      description_md:
        type: string
      endpoint_id:
        type: integer
      external_url:
        type: string
      host_id:
        type: integer
      poc:
        type: string
      reward_amount:
        type: number
      reward_currency:
        type: string
      service_id:
        type: integer
      status:
        type: string
      target:
        type: string
      title:
        type: string
      url:
        type: string
      vrt:
        type: string
    type: object
...
  sql.NullFloat64:
    properties:
      float64:
        format: float64
        type: number
      valid:
        description: Valid is true if Float64 is not NULL
        type: boolean
    type: object
  sql.NullInt64:
    properties:
      int64:
        format: int64
        type: integer
      valid:
        description: Valid is true if Int64 is not NULL
        type: boolean
    type: object
info:
  contact: {}
  description: Bug bounty & continuous recon backend.
  title: Falcon API
  version: 0.1.0
paths:
...
  /api/programs/{id}/reports:
    get:
      parameters:
      - description: program id
        in: path
        name: id
        required: true
        type: integer
      - description: page
        in: query
        name: page
        type: integer
      - description: page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.pageResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.errorResp'
      security:
      - BearerAuth: []
      summary: List reports
      tags:
      - reports
    post:
      consumes:
      - application/json
      parameters:
      - description: program id
        in: path
        name: id
        required: true
        type: integer
      - description: report body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.reportReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_erhan_falcon_api_internal_models.Report'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.errorResp'
      security:
      - BearerAuth: []
      summary: Create report
      tags:
      - reports
  /api/programs/{id}/runs:
    delete:
      parameters:
      - description: program id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.errorResp'
      security:
      - BearerAuth: []
      summary: Delete all runs for a program
      tags:
      - runs
    get:
      parameters:
      - description: program id
        in: path
        name: id
        required: true
        type: integer
      - description: page
        in: query
        name: page
        type: integer
      - description: page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.pageResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.errorResp'
      security:
      - BearerAuth: []
      summary: List runs
      tags:
      - runs
...
  /api/reports/{id}:
    delete:
      parameters:
      - description: report id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.errorResp'
      security:
      - BearerAuth: []
      summary: Delete report
      tags:
      - reports
    get:
      parameters:
      - description: report id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_erhan_falcon_api_internal_models.Report'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.errorResp'
      security:
      - BearerAuth: []
      summary: Get report
      tags:
      - reports
    patch:
      consumes:
      - application/json
      parameters:
      - description: report id
        in: path
        name: id
        required: true
        type: integer
      - description: fields
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.reportReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_erhan_falcon_api_internal_models.Report'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.errorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.errorResp'
      security:
      - BearerAuth: []
      summary: Update report
      tags:
      - reports
  /api/runs/{id}:
    delete:
      parameters:
      - description: run id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.errorResp'
      security:
      - BearerAuth: []
      summary: Delete run
      tags:
      - runs
    get:
      parameters:
      - description: run id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.errorResp'
      security:
      - BearerAuth: []
      summary: Get run with steps
      tags:
      - runs
...
securityDefinitions:
  BearerAuth:
    description: 'JWT token issued by /api/auth/login. Format: "Bearer {token}".'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"